sensor_client/requirements.txt
fastapi
pymysql
uvicorn
pyserial
requests
